Correlate should be finished for t2, but t3 has not been tested. Start-stop mode has also been implemented.

+++ src/photon.h
@@ -25,8 +25,8 @@ int photon_queue_pop(photon_queue_t *queue, void *photon);
 int photon_queue_push(photon_queue_t *queue, void *photon);
 int photon_queue_front(photon_queue_t const *queue, void *photon);
 int photon_queue_back(photon_queue_t const *queue, void *photon);
-size_t photon_queue_index(photon_queue_t const *queue, void *photon, 
-		int const index);
+int photon_queue_index(photon_queue_t const *queue, void *photon, 
+		size_t const index);
 size_t photon_queue_size(photon_queue_t const *queue);
 void photon_queue_sort(photon_queue_t *queue);
